+++ data/cutscenes.lua
@@ -32,6 +32,14 @@ cutscenes.tutorial_start = Cutscene:new("tutorial_end", {
 
         duration = 5.0,
         enter = function(scene)
+            game.game_ui:start_title("Ninesliced", "", "{menu.credits.game_by}", 0.5, 4.0, 0.5)
+        end,
+    }),
+    CutsceneScene:new({
+        description = "",
+
+        duration = 5.0,
+        enter = function(scene)
             game.game_ui:start_title("Léo Bernard", "Yolwoocle", "{menu.credits.game_by}", 0.5, 4.0, 0.5)
         end,
     }),
@@ -46,17 +54,9 @@ cutscenes.tutorial_start = Cutscene:new("tutorial_end", {
     CutsceneScene:new({
         description = "",
 
-        duration = 5.0,
-        enter = function(scene)
-            game.game_ui:start_title("Martin Domergue", "Verbaudet", "{menu.credits.sound_design}", 0.5, 4.0, 0.5)
-        end,
-    }),
-    CutsceneScene:new({
-        description = "",
-
         duration = 6.0,
         enter = function(scene)
-            game.game_ui:start_title("Ninesliced", "", "{menu.credits.published_by}", 0.5, 4.0, 0.5)
+            game.game_ui:start_title("Martin Domergue", "Verbaudet", "{menu.credits.sound_design}", 0.5, 4.0, 0.5)
         end,
     }),
     CutsceneScene:new({