+++ scripts/ui/game_ui.lua
@@ -468,12 +468,17 @@ function GameUI:draw_fury()
 	local w = math.min(self.fury_visual_width, math.huge)
 	local h = 8
 	-- local col = ternary(self.fury_is_flashing, COL_YELLOW_ORANGE, COL_ORANGE)
-	local col = COL_YELLOW_ORANGE
+
+	local col, text_col = COL_YELLOW_ORANGE, COL_LIGHT_YELLOW
+	if game.level.has_energy_drink then
+		col, text_col = COL_MID_BLUE, COL_LIGHT_BLUE
+	end
+
 	rect_color(col, "fill", CANVAS_WIDTH/2 - w, y-h/2, w*2, h)
 	print_wavy_centered_outline_text(
-		COL_LIGHT_YELLOW, nil, 
-		-- Text:text("game.combo", game.level.fury_combo), 
-		concat(game.level.fury_combo), 
+		text_col, nil, 
+		Text:text("game.combo", game.level.fury_combo), 
+		-- concat(game.level.fury_combo), 
 		CANVAS_WIDTH/2, y-h/2, nil, self.t, 2, 9, 0.8
 	)